Se añadio seguridad para endpoint usuario

+++ src/usuario/usuario.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Put, Request, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, NotFoundException, Param, Post, Put, Request, UseGuards } from '@nestjs/common';
 import { UsuarioService } from './usuario.service';
 import { Usuario } from './usuario.entity';
 import { JwtAuthguard } from 'src/auth/jwt-auth.guard';
@@ -9,15 +9,25 @@ export class UsuarioController {
 
   
   @Get()
-  async findAll(): Promise<Usuario[]> {
+  @UseGuards(JwtAuthguard)
+  async findAll(@Request() request): Promise<Usuario[]> {
+    if (!request.user.admin) {
+      throw new ForbiddenException('No tienes permiso');
+    }
     return await this.usuarioService.findAll();
   }
 
   @Get(':id')
-  async findOne(@Param('id') id:number): Promise<Usuario>{
+  @UseGuards(JwtAuthguard)
+  async findOne(@Param('id') id: number, @Request() request): Promise<Usuario> {
+    //si no sos admin ni el dueño del id, no puedes acceder
+    if (!request.user.admin && request.user.id !== id) {
+      throw new ForbiddenException('No tienes permiso para ver este usuario');
+    }
+
     const usuario = await this.usuarioService.findOne(id);
-    if(!usuario){
-      throw new NotFoundException(`Usuario con ID ${id} no encontrada`);
+    if (!usuario) {
+      throw new NotFoundException(`Usuario con ID ${id} no encontrado`);
     }
     return usuario;
   }
@@ -27,11 +37,22 @@ export class UsuarioController {
     return await this.usuarioService.create(usuarioData);
   }
 
+  @UseGuards(JwtAuthguard)
   @Put(':id')
   async update(
     @Param('id') id: number,
-    @Body() usuarioData: Partial<Usuario>
+    @Body() usuarioData: Partial<Usuario>,
+    @Request() request
   ): Promise<Usuario> {
+    //bloqueo para no editar otros
+    if (!request.user.admin && request.user.id !== id) {
+      throw new ForbiddenException('No tienes permiso para actualizar este usuario');
+    }
+    //no permitir volverse admin
+    if (!request.user.admin) {
+      delete usuarioData.admin;
+    }
+
     const updated = await this.usuarioService.update(id, usuarioData);
     if (!updated) {
       throw new NotFoundException(`No se pudo actualizar la usuario con ID ${id}`);
@@ -39,8 +60,13 @@ export class UsuarioController {
     return updated;
   }
 
+  @UseGuards(JwtAuthguard)
   @Delete(':id')
-  async remove(@Param('id') id: number): Promise<{ message: string }> {
+  async remove(@Param('id') id: number, @Request() request): Promise<{ message: string }> {
+    //bloqueo para no editar otros
+    if (!request.user.admin && request.user.id !== id) {
+      throw new ForbiddenException('No tienes permiso para borrar este usuario');
+    }
     await this.usuarioService.remove(id);
     return { message: `Usuario con ID ${id} eliminada correctamente` };
   }
